refactor: changing generator Type to custom types
