+++ customTypes/types.go
@@ -25,3 +25,24 @@ func (l Language) String() string {
 		return "unknown"
 	}
 }
+
+type GeneratorType int
+
+const (
+	WifiGeneratorType GeneratorType = iota
+	RandomGeneratorType
+	TextInputGeneratorType
+)
+
+func (g GeneratorType) String() string {
+	switch g {
+	case RandomGeneratorType:
+		return "random"
+	case WifiGeneratorType:
+		return "wifi"
+	case TextInputGeneratorType:
+		return "textInput"
+	default:
+		return "wifi"
+	}
+}
